refactor engine and model agent
