+++ lmdeploy/pytorch/engine/executor/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+def build_executor():
+    """build model agent executor."""
