+++ lmdeploy/pytorch/engine/executor/base.py
@@ -1,0 +1,6 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+class ExecutorBase:
+
+    def build_model(self):
+        """build model."""
+        raise NotImplementedError('Not Implemented.')
