+++ lmdeploy/pytorch/engine/model_agent.py
@@ -1,11 +1,11 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import asyncio
-from typing import Dict
+from typing import Any, Dict
 
 import torch
 import torch.distributed as dist
 
-from lmdeploy.utils import get_logger
+from lmdeploy.utils import get_logger, logging_timer
 
 from ..backends import get_backend
 from ..config import BackendConfig, CacheConfig, ModelConfig
@@ -15,6 +15,8 @@ from ..models.patch import add_adapters, build_patched_model, update_custom_modu
 from ..utils import get_gpu_memory
 from ..weight_loader.model_weight_loader import load_model_weights
 from .cache_engine import CacheEngine
+from .dist_proc_manager import broadcast_inputs
+from .logits_process import FusedLogitsProcessor, SamplingInputs
 
 logger = get_logger('lmdeploy')
 
@@ -180,15 +182,28 @@ def model_forward(
     return dict(hidden_states=output, model_metas=model_metas)
 
 
+def _batch_stopping_criteria(token_ids: torch.Tensor, stop_words: torch.Tensor, num_appendable_ids: torch.Tensor):
+    """batched stopping criteria."""
+    num_appendable_ids = num_appendable_ids - 1
+    # one more step to cache last token(stop word)
+    stopped = num_appendable_ids < 0
+    if stop_words is not None:
+        sw_stopped = (token_ids[:, None] == stop_words).any(1)
+        one_ids = torch.clamp_max(num_appendable_ids, 0)
+        num_appendable_ids = torch.where(sw_stopped, one_ids, num_appendable_ids)
+    return stopped, num_appendable_ids
+
+
 SwapMap = Dict[int, int]
 
 
 class AutoModelAgent:
     """Base model agent."""
 
-    def __init__(self, model_config: ModelConfig, cache_config: CacheConfig):
+    def __init__(self, model_config: ModelConfig, cache_config: CacheConfig, tokenizer: Any):
         self.model_config = model_config
         self.cache_config = cache_config
+        self.tokenizer = tokenizer
 
     async def async_forward(self, inputs: ModelInputs, swap_in_map: SwapMap, swap_out_map: SwapMap):
         """model forward.
@@ -208,6 +223,189 @@ class AutoModelAgent:
         """get input processor."""
         raise NotImplementedError('Not implemented.')
 
+    async def _async_model_forward(self, inputs: ModelInputs, swap_in_map: Dict, swap_out_map: Dict,
+                                   return_logits: bool):
+        """model forward."""
+        max_prefill_token_num = self.cache_config.max_prefill_token_num
+        swap_done = False
+
+        class _OutputGather:
+            """output gather."""
+
+            def __init__(self, max_seq_len):
+                self._max_seq_len = max_seq_len
+                self._start = 0
+                self._output = None
+
+            def gather(self, output):
+                """gather."""
+                tmp_output = output['hidden_states']
+
+                if not return_logits:
+                    self._output = tmp_output
+                    return
+
+                out_logits = self._output
+                start = self._start
+                seq_len = tmp_output.size(-2)
+                if out_logits is None:
+                    out_logits = tmp_output.new_empty(1, self._max_seq_len, tmp_output.size(-1), device='cpu')
+                out_logits[:, start:start + seq_len].copy_(tmp_output, non_blocking=True)
+                self._start = start + seq_len
+                self._output = out_logits
+
+            def get_output(self):
+                """get tmp_output."""
+                if not return_logits:
+                    return self._output[:, -1:]
+                torch.cuda.synchronize()
+                return self._output
+
+        async def __forward(inputs):
+            """forward."""
+            nonlocal swap_done, swap_in_map, swap_out_map
+            if swap_done:
+                return await self.async_forward(inputs, swap_in_map=dict(), swap_out_map=dict())
+            else:
+                swap_done = True
+                return await self.async_forward(inputs, swap_in_map=swap_in_map, swap_out_map=swap_out_map)
+
+        async def __long_context_single_forward(inputs):
+            """one large sequence."""
+            seq_len = inputs.seq_length
+            max_seq_len = inputs.seq_length[0]
+            batch_size = seq_len.size(0)
+            assert batch_size == 1
+
+            new_inputs = inputs.split(max_prefill_token_num)
+
+            model_metas = new_inputs[0].model_metas
+            output_gather = _OutputGather(max_seq_len)
+            for inp in new_inputs:
+                inp.model_metas = model_metas
+                tmp_out = await __forward(inp)
+                model_metas = tmp_out.get('model_metas')
+                output_gather.gather(tmp_out)
+                tmp_out.pop('hidden_states', None)
+            tmp_out['hidden_states'] = output_gather.get_output()
+            return tmp_out
+
+        if inputs.input_ids.numel() <= max_prefill_token_num:
+            ret = await __forward(inputs)
+            if not return_logits and not inputs.is_decoding:
+                last_token_loc = inputs.seq_length.cumsum(0) - 1
+                ret['hidden_states'] = ret['hidden_states'][:, last_token_loc]
+        else:
+            ret = await __long_context_single_forward(inputs)
+            if not return_logits and not inputs.is_decoding:
+                last_token_loc = [-1]
+                ret['hidden_states'] = ret['hidden_states'][:, last_token_loc]
+            else:
+                ret['hidden_states'] = ret['hidden_states'].to('cuda')
+
+        hidden_states = ret.pop('hidden_states')
+        logits = self.get_logits(hidden_states)
+        ret['logits'] = logits
+        return ret
+
+    @logging_timer('SamplingLogits', logger)
+    async def async_sampling_logits(self, logits: torch.Tensor, all_ids: torch.Tensor, guided_input_ids: torch.Tensor,
+                                    sampling_inputs: SamplingInputs, inputs: ModelInputs, ignore_eos: torch.Tensor):
+        """sampling logits."""
+
+        def __get_last_logits():
+            """get last logits."""
+            seq_length = inputs.seq_length
+            if len(seq_length) == logits.size(0):
+                return logits
+
+            last_idx = seq_length.cumsum(-1) - 1
+            return logits[last_idx, :]
+
+        split_logits = __get_last_logits()
+        logits_processor = FusedLogitsProcessor(sampling_inputs, ignore_eos, self.tokenizer)
+        logits = await logits_processor(all_ids, guided_input_ids, split_logits)
+        next_token_ids = logits_processor.sampling(logits)
+
+        return next_token_ids
+
+    async def _async_step_background(self, inputs: ModelInputs, swap_in_map: Dict, swap_out_map: Dict,
+                                     all_ids: torch.Tensor, guided_input_ids: torch.Tensor,
+                                     sampling_inputs: SamplingInputs, num_appendable_ids: torch.LongTensor,
+                                     num_ignore_eos: torch.LongTensor, loop_count: int, return_logits: bool,
+                                     output_que: asyncio.Queue):
+        """asyc forward task."""
+
+        def __update_inputs(next_token_ids):
+            """update inputs."""
+            nonlocal all_ids, guided_input_ids
+            inputs.update(next_token_ids)
+            if all_ids is not None:
+                all_ids = torch.cat([all_ids, next_token_ids[:, None].to(all_ids.device)], 1)
+            if guided_input_ids is not None:
+                guided_input_ids = torch.cat([guided_input_ids, next_token_ids[:, None].to(guided_input_ids.device)], 1)
+            if sampling_inputs.random_offsets is not None:
+                sampling_inputs.random_offsets += 1
+
+        logger.debug('<ForwardTask>: '
+                     f'batch_size={inputs.seq_length.size(0)} '
+                     f'num_tokens={inputs.input_ids.size(-1)}')
+        inputs = inputs.to_device('cuda')
+        is_decoding = inputs.is_decoding
+        if all_ids is not None:
+            all_ids = all_ids.cuda()
+        if guided_input_ids is not None:
+            guided_input_ids = guided_input_ids.cuda()
+        sampling_inputs = sampling_inputs.to_device('cuda')
+        num_appendable_ids = num_appendable_ids.cuda()
+        num_ignore_eos = num_ignore_eos.cuda()
+
+        # dist tools
+        dist_ctx = get_dist_manager().current_context()
+
+        for idx in range(loop_count):
+
+            # broadcast inputs
+            if dist_ctx.tp > 1:
+                tp_cpu_group = dist_ctx.tp_cpu_group
+                (inputs, swap_in_map, swap_out_map) = broadcast_inputs(0, [inputs, swap_in_map, swap_out_map],
+                                                                       tp_cpu_group)
+
+            # inference
+            output = await self._async_model_forward(inputs,
+                                                     swap_in_map=swap_in_map,
+                                                     swap_out_map=swap_out_map,
+                                                     return_logits=return_logits)
+            logits = output['logits']
+            logits = logits[0]  # [bs, seq, prob] -> [seq, prob]
+
+            # sampling
+            next_token_ids = await self.async_sampling_logits(logits, all_ids, guided_input_ids, sampling_inputs,
+                                                              inputs, num_ignore_eos > 0)
+            num_ignore_eos = num_ignore_eos - 1
+
+            # stopping criteria
+            stopped, num_appendable_ids = _batch_stopping_criteria(next_token_ids, sampling_inputs.stop_words,
+                                                                   num_appendable_ids)
+
+            # send output
+            model_metas = output.get('model_metas')
+            event = torch.cuda.Event()
+            event.record()
+            output = dict(next_token_ids=next_token_ids,
+                          logits=logits,
+                          stopped=stopped,
+                          model_metas=model_metas,
+                          event=event)
+            output_que.put_nowait(output)
+
+            # update for next loop
+            if is_decoding and idx < loop_count - 1:
+                swap_in_map = dict()
+                swap_out_map = dict()
+                inputs.model_metas = model_metas
+                __update_inputs(next_token_ids)
+
 
 class BaseModelAgent(AutoModelAgent):
     """Base model agent.
@@ -226,8 +424,9 @@ class BaseModelAgent(AutoModelAgent):
                  model_config: ModelConfig,
                  cache_config: CacheConfig,
                  backend_config: BackendConfig,
+                 tokenizer: Any,
                  adapters: Dict[str, str] = None):
-        super().__init__(model_config=model_config, cache_config=cache_config)
+        super().__init__(model_config=model_config, cache_config=cache_config, tokenizer=tokenizer)
         device = 'cuda'
         self.backend_config = backend_config
         dist_ctx = get_dist_manager().current_context()
@@ -313,6 +512,7 @@ class BaseModelAgent(AutoModelAgent):
 def build_model_agent(model_path: str,
                       cache_config: CacheConfig,
                       backend_config: BackendConfig,
+                      tokenizer: Any,
                       trust_remote_code: bool,
                       adapters: Dict[str, str] = None,
                       tp: int = 1,
@@ -337,5 +537,6 @@ def build_model_agent(model_path: str,
                                  model_config=model_config,
                                  cache_config=cache_config,
                                  backend_config=backend_config,
+                                 tokenizer=tokenizer,
                                  adapters=adapters)
     return model_agent
